requirements.txt
spacy
textblob
pyttsx3
pypdf
chromadb
langchain_groq
langchain_community
langchain_core
sentence_transformers
en-core-web-sm@https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz

